Include a Python check in the install.sh file, and also modularize the setup process.
